refactor: separate static vs instance ImageUtils; breakfown files further; move constants/configs to better places;
fix: refactor; move instance ops to core; minor changes

refactor: extract tuning config from instance

fix: add screeninfo for simple window info
