+++ requirements.txt
@@ -1,7 +1,8 @@
-matplotlib>=3.3.1
-numpy>=1.16.1
-pandas>=1.1.1
-deepmerge>=0.1.0
-dotmap>=1.3.17
-jsonschema>=4.3.3
+deepmerge>=1.0.1
+dotmap>=1.3.30
+jsonschema>=4.16.0
+matplotlib>=3.5.3
+numpy>=1.23.2
+pandas>=1.4.3
 rich>=12.6.0
+screeninfo>=0.8.1
