+++ samples/sample5/template.json
@@ -1,7 +1,5 @@
 {
   "Globals": {
-    "window_width": 1189,
-    "window_height": 1682,
     "display_width": 1189,
     "display_height": 1682
   },
