+++ samples/sample6/template.json
@@ -1,7 +1,5 @@
 {
   "Globals": {
-    "window_width": 2480,
-    "window_height": 3508,
     "display_width": 2480,
     "display_height": 3508
   },
