+++ src/defaults/__init__.py
@@ -1,4 +1,6 @@
 # https://docs.python.org/3/tutorial/modules.html#:~:text=The%20__init__.py,on%20the%20module%20search%20path.
-from .config import *  # NOQA
-from .evaluation import *  # NOQA
-from .template import *  # NOQA
+# Use all imports relative to root directory
+# (https://chrisyeh96.github.io/2017/08/08/definitive-guide-python-imports.html)
+from src.defaults.config import *  # NOQA
+from src.defaults.evaluation import *  # NOQA
+from src.defaults.template import *  # NOQA
